testsuite: modify t3301-system-latestart.t for flux-sched
Problem: t3301-system-latestart.t does not load Fluxion modules
and TEST_UNDER_FLUX_AUGMENT_R is required to avoid a hand.

Add environment variables necessary to load fluxion modules
in the test. Add a check that the correct module is loaded,
and a test for `flux resource list` as a debug step.