testsuite: fix expiration propagation test
Problem: The test in t4011-match-duration.t tests that the duration
of a job is inherited from the enclosing instance duration, but we
should really be testing that the *expiration* of a child job is
inherited from the parent when no duration is specified. O/w, a job
submitted when the instance has only a few minutes remaining could
have its expiration set for long after the instance is terminated.

Update the test to ensure that expiration of a job with no duration
specified is inherited from the enclosing instance expiration.
Use flux-alloc(1) to launch the test instance instead of running
standalone flux-start(1) instances even though test_under_flux()
is used.