testsuite: add test for expiration update inheritance
Problem: No tests in the testsuite ensure that an expiration update
of a job containing a Flux instance is inherited by the subinstance
Fluxion scheduler.

Add a test to t4011-match-duration.t that launches a subinstance,
extends the duration of that instance, then ensures that new jobs
submitted to the instance inherit the updated expiration.