resource: support `shrink`in RFC 28 resource acquisition response
Problem: RFC 28 defines a `shrink` key in resource acquisition response
which, when set, indicates that resources have been permanently lost
and should not be considered for satisfiability of future jobs.

Add a `shrink()` method to the dfu traverser which supports finding
subgraphs by rank ids and removes them. Call this function with the set
of `shrink` ranks if they appear in a resource acquisition response.