resource: add partial cancel by rank set
Problem: When a set of ranks are included in a `shrink` key in the
resource acquisition response, there should first be a partial cancel
operation performed on any jobs with resources on the included ranks,
but this is not currently done.

Add a new function partial_cancel_by_ranks() which takes a set of ranks,
iterates over all jobids assigned to those ranks and calls remove()
for each one using a skeleton Rv1 object.

Note: this is known not to work at this time, so errors are currently
logged, but otherwise ignored.